model/requirements.txt
pydantic-settings~=2.8.1
python-dotenv~=1.1.0
elasticsearch~=8.17.2
sqlmodel~=0.0.24
fastapi~=0.115.12
pandas~=2.2.3
torch~=2.6.0
transformers~=4.50.3
numpy~=2.2.4
scikit-learn~=1.6.1
stable-baselines3~=2.6.0
gym~=0.26.2
psutil~=5.9.0